Fix Indent and Remove Let It Go function :snowman:
Now, I really think it's awesome.

I Figured out that LetItGo is actually the function In it self.

So, if Transform is defined. It returns a new In() with the transformed
data.

If transform is NOT defined. It returns a new In() with the current data

And now we can do stuff like

new In(1)(log)(add(2))(log) // output 3

Discussed here
https://github.com/felquis/inout/commit/fe00815cf7ffea1762501a3f36bf5783
7eca2fe9#commitcomment-12726247