Add add Transform
Now it's kind of cool, a Transform can return a Transform.

This function first set the `num` variable, it is the number that will
be add to the data.

you can test it in this way

add(1)

If you took attention, it returned a new function that receive a data
parameter.

Yes, it means that you can use this module like this.

add(1)(2)

Or with In

In(2)(add(2))(log) // Output 4