Out with no Transform returns the closure data
After this change, you could do this:
new In(1)(log)()()()()(log)

The four calls between `log` always return the previous data. I thought
it was good cause there was no way to expose the data to a variable,
you could only get the data with Side Effects.

```js
// A function to cause a side effect in the variable `result`
function expose (data) {
  result = data;
  return data;
}

// The given variable
var result;

// Use the module
In(1)(expose);

log(expose) // 1

```

But I don't like Side Effects, implement side effects probably was
awesome to the developer who made it, but it is not headache proof.
But, developers couldn't do ()()()() even if they want to. It's not
elegant. But I believe this tinny code will be used with other codes
around it. And sometimes you just want to use a few modules and pass
the result to another framework or library to deal with the result. So,
we need an elegant way to get the output data.

Just like you tie the shoelaces of your old all star. You can get the
Data by calling the Out function with no Transform.

```js
var result = In(1)()
```

Remember that `In` returns `Out`, if you call `Out` with a `Transform`,
it will call `In` with the Transformed data, which will return a new
`Out` function. But, if you call Out with no Transform. It'll output
the data.

And now `result` is 1, the value passed to In.

```js
var result = In(1)(add(2))()
```

Can you guess the value of result? I hope so :)

My next commit will be an update of `inin.js` example, check it out.